# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para memorias de cálculo de instalaciones eléctricas según normativa NOM (México)",
        "title": "Garfex Calculadora de Filtros API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/calculos/caida-tension": {
            "post": {
                "description": "Calcula la caída de tensión en un circuito según IEEE-141. Requiere factor_potencia (cosθ) en rango (0, 1].",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Caída de Tensión"
                ],
                "summary": "Calcular caída de tensión",
                "parameters": [
                    {
                        "description": "Datos del circuito y conductor",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CaidaTensionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Caída de tensión calculada",
                        "schema": {
                            "$ref": "#/definitions/http.CaidaTensionResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación o datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/http.CaidaTensionResponseError"
                        }
                    },
                    "422": {
                        "description": "No se encontró la impedancia para el calibre y canalización",
                        "schema": {
                            "$ref": "#/definitions/http.CaidaTensionResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.CaidaTensionResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ResultadoCaidaTension": {
            "type": "object",
            "properties": {
                "caida_volts": {
                    "type": "number"
                },
                "cumple": {
                    "type": "boolean"
                },
                "impedancia": {
                    "type": "number"
                },
                "limite_porcentaje": {
                    "type": "number"
                },
                "porcentaje": {
                    "type": "number"
                },
                "reactancia": {
                    "type": "number"
                },
                "resistencia": {
                    "type": "number"
                }
            }
        },
...
        "http.CaidaTensionRequest": {
            "type": "object",
            "required": [
                "calibre",
                "corriente_nominal",
                "factor_potencia",
                "hilos_por_fase",
                "limite_caida",
                "longitud_circuito",
                "material",
                "sistema_electrico",
                "tension",
                "tipo_canalizacion",
                "tipo_voltaje"
            ],
            "properties": {
                "calibre": {
                    "description": "Calibre es el calibre del conductor (ej: \"2 AWG\", \"250 MCM\")",
                    "type": "string"
                },
                "corriente_nominal": {
                    "description": "CorrienteNominal es la corriente nominal del circuito en amperes",
                    "type": "number"
                },
                "factor_potencia": {
                    "description": "FactorPotencia es el coseno del ángulo de desfasamiento (cosθ)\nRepresenta el factor de potencia del circuito: 0 \u003c FP ≤ 1\nEjemplos: 0.85 (motores típicos), 0.9 (cargas mixtas), 1.0 (carga resistiva pura)",
                    "type": "number",
                    "maximum": 1
                },
                "hilos_por_fase": {
                    "description": "HilosPorFase es el número de hilos por fase",
                    "type": "integer",
                    "minimum": 1
                },
                "limite_caida": {
                    "description": "LimiteCaida es el límite de caída de tensión en porcentaje",
                    "type": "number"
                },
                "longitud_circuito": {
                    "description": "LongitudCircuito es la longitud del circuito en metros",
                    "type": "number"
                },
                "material": {
                    "description": "Material es el material del conductor (\"Cu\" = cobre, \"Al\" = aluminio)",
                    "type": "string"
                },
                "sistema_electrico": {
                    "description": "SistemaElectrico es el tipo de sistema eléctrico\nValores: \"MONOFASICO\", \"BIFASICO\", \"DELTA\", \"ESTRELLA\"",
                    "type": "string"
                },
                "tension": {
                    "description": "Tension es la tensión del sistema en volts",
                    "type": "integer"
                },
                "tipo_canalizacion": {
                    "description": "TipoCanalizacion es el tipo de canalización\nValores: \"TUBERIA_PVC\", \"TUBERIA_METALICA\", \"CHAROLA\"",
                    "type": "string"
                },
                "tipo_voltaje": {
                    "description": "TipoVoltaje indica si el voltaje ingresado es fase-neutro o fase-fase\nValores: \"FASE_NEUTRO\" (Vfn), \"FASE_FASE\" (Vff), también acepta \"FN\" o \"FF\"\nEjemplos: 127V es típicamente Vfn, 220V es típicamente Vff",
                    "type": "string"
                }
            }
        },
        "http.CaidaTensionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ResultadoCaidaTension"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CaidaTensionResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}